+++ packages/agents-extensions/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @openai/agents-extensions
 
+## 0.5.1
+
+### Patch Changes
+
+- @openai/agents@0.5.1
+- @openai/agents-core@0.5.1
+
+## 0.5.0
+
+### Patch Changes
+
+- Updated dependencies [c590057]
+  - @openai/agents@0.5.0
+  - @openai/agents-core@0.5.0
+
+## 0.4.15
+
+### Patch Changes
+
+- Updated dependencies [40c1709]
+  - @openai/agents-core@0.4.15
+  - @openai/agents@0.4.15
+
 ## 0.4.14
 
 ### Patch Changes